requirements.txt
gunicorn
dj_database_url
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1.1
pytz==2019.1
sqlparse==0.3.0
psycopg2
psycopg2-binary
whitenoise
